+++ utoipa-gen/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Add support for `std::ops::{Range, RangeTo, RangeFull}` (https://github.com/juhaku/utoipa/pull/1354)
+* Add support for `std::ops::{Range, RangeFrom, RangeTo}` (https://github.com/juhaku/utoipa/pull/1354)
 
 ## 5.3.1 - Jan 6 2025