+++ utoipa/CHANGELOG.md
@@ -5,7 +5,7 @@ to look into changes introduced to **`utoipa-gen`**.
 
 ## Unreleased
 
-* Add support for `std::ops::{Range, RangeTo, RangeFull}` (https://github.com/juhaku/utoipa/pull/1354)
+* Add support for `std::ops::{Range, RangeFrom, RangeTo}` (https://github.com/juhaku/utoipa/pull/1354)
 
 ### Changed